feat: refactor organization and candidate models to import Status and Country from shared module
